PR binutils/4292 	* nm.c (print_value): Cast bfd_vma to unsigned long to suit 	value_format_32bit.

+++ binutils/nm.c
@@ -1334,7 +1334,7 @@ print_value (bfd *abfd ATTRIBUTE_UNUSED, bfd_vma val)
   switch (print_width)
     {
     case 32:
-      printf (value_format_32bit, val);
+      printf (value_format_32bit, (unsigned long) val);
       break;
 
     case 64:
